requirements.txt
dash==2.14.2
dash_bootstrap_components==1.5.0
numpy==1.26.2
pandas==2.1.3
plotly==5.18.0
pytest==7.4.3
tabulate==0.9.0
dash-extensions==1.0.4
lxml==4.9.3
openpyxl==3.1.2
orjson==3.9.10
pyyaml==6.0.1
typing_extensions>=4.1.0
mypy_extensions>=1.0.0
pyuwsgi==2.0.23
scipy==1.11.4
flask_httpauth==4.8.0
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
opentelemetry-sdk==1.21.0
opentelemetry-api==1.21.0
opentelemetry-distro==0.42b0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-exporter-otlp==1.21.0
Pillow==9.5.0
cairosvg==2.7.1
Levenshtein==0.23.0
rapidfuzz==3.5.2
beautifulsoup4==4.12.2
apscheduler==3.10.4
ratelimit==2.2.1
cachetools==5.3.2